chore: move config to util folder
