+++ apps/mech/common-util/hooks/useHandleRoute.ts
@@ -4,7 +4,7 @@ import { useCallback, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 
 import { setChainId } from 'store/setup/actions';
-import { ALL_SUPPORTED_CHAINS, FIRST_SUPPORTED_CHAIN } from 'components/Login/config';
+import { ALL_SUPPORTED_CHAINS, FIRST_SUPPORTED_CHAIN } from 'common-util/login/config';
 import { PAGES_TO_LOAD_WITHOUT_CHAIN_ID, URL } from 'util/constants';
 
 const isValidNetworkName = (name: string) => {
