+++ apps/mech/components/Layout/index.tsx
@@ -6,7 +6,7 @@ import { useSelector } from 'react-redux';
 
 import { COLOR } from '@autonolas/frontend-library';
 
-import { ALL_SUPPORTED_CHAINS } from 'components/Login/config';
+import { ALL_SUPPORTED_CHAINS } from 'common-util/login/config';
 import { getSupportedNetworks } from 'common-util/functions';
 import { useHandleRoute } from 'common-util/hooks/useHandleRoute';
 import { PAGES_TO_LOAD_WITH_CHAIN_ID } from 'util/constants';
