+++ apps/mech/w3m-button.d.ts
@@ -1,0 +1,11 @@
+declare namespace JSX {
+  interface IntrinsicElements {
+    'w3m-button': {
+      disabled?: boolean;
+      balance?: 'show' | 'hide';
+      size?: 'md' | 'sm';
+      label?: string;
+      loadingLabel?: string;
+    };
+  }
+}
